refactor: extract shared types to lib/types.ts
